files moved to personList from people, import updated, reducers names changed
