+++ src/core/saga.js
@@ -1,7 +1,7 @@
 import { all } from "redux-saga/effects";
 import { movieSaga } from "../features/movies/movieSaga";
-import { peopleSaga } from "../features/people/peopleSaga";
 import { detailsSaga } from "../features/people/personDetails/detailsSaga";
+import { peopleSaga } from "../features/people/personList/peopleSaga";
 
 export default function* saga() {
   yield all([
